# comsapi/app/src/docs/v1.api-spec.yaml
...
openapi: 3.0.2
info:
  version: 1.0.0
  title: Common Object Management Service (COMS)
  description: A microservice for managing access control to S3 Objects
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  contact:
    name: NR Common Service Showcase
    email: NR.CommonServiceShowcase@gov.bc.ca
servers:
  - url: /api/v1
    description: This Server
security:
  - BasicAuth: []
  - BearerAuth: []
    OpenID: []
...
paths:
...
  /bucket/{bucketId}/sync:
    get:
      summary: Synchronize a bucket
      description: >-
        Synchronize the contents of an existing bucket with COMS so that it can
        track and manage the objects in it. This avoids the need to re-upload
        preexisting files through the COMS API. This endpoint does not guarantee
        immediately synchronized results. Synchronization latency will be
        affected by the remaining number of objects awaiting synchronization.
      tags:
        - Sync
      operationId: syncBucket
      parameters:
        - $ref: '#/components/parameters/Path-BucketId'
      responses:
        '202':
          $ref: '#/components/responses/AddedQueueLength'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Error'
...
  /object:
    put:
      summary: Creates an object
      description: >-
        Creates an object in the bucket specified with bucketId parameter or
        default bucket configured. If COMS is running in either 'OIDC' or 'Full'
        mode, any objects created with OIDC user authentication will have all
        object permissions assigned to them by default. This endpoint will do a
        best attempt effort to create the object as requested as long as it does
        not already exist in the destination bucket and path. Existing objects
        should be managed with the updateObject endpoint instead.
      operationId: createObject
      tags:
        - Object
      parameters:
        - $ref: '#/components/parameters/Header-ContentDisposition'
        - $ref: '#/components/parameters/Header-ContentLength'
        - $ref: '#/components/parameters/Header-ContentType'
        - $ref: '#/components/parameters/Header-Metadata'
        - $ref: '#/components/parameters/Query-TagSet'
        - in: query
          name: bucketId
          description: Uuid representing the bucket
          schema:
            type: string
            format: uuid
            example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
      requestBody:
        description: Raw binary representation of the file
        required: true
        content:
          '*/*':
            schema:
              type: string
              description: >-
                Contains the binary representation of the file to upload.
                Maximum filesize of 5TB.
              format: binary
              maxLength: 5497558138880
      responses:
        '200':
          description: Returns the created object data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response-ObjectDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Conflict. Bucket already contains the object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response-Conflict'
        '411':
          $ref: '#/components/responses/LengthRequired'
        '413':
          $ref: '#/components/responses/ContentTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Error'
    get:
      summary: Search for objects
      description: >-
        Returns a list of objects matching all search criteria across all known
        versions of objects. Search criteria on string attributes will match on
        partial results and ignore case sensitivity.
      operationId: searchObjects
      tags:
        - Object
      parameters:
        - $ref: '#/components/parameters/Header-Metadata'
        - $ref: '#/components/parameters/Query-ObjectId'
        - $ref: '#/components/parameters/Query-BucketId'
        - $ref: '#/components/parameters/Query-Path'
        - $ref: '#/components/parameters/Query-Active'
        - $ref: '#/components/parameters/Query-DeleteMarker'
        - $ref: '#/components/parameters/Query-Latest'
        - $ref: '#/components/parameters/Query-Public'
        - $ref: '#/components/parameters/Query-MimeType'
        - $ref: '#/components/parameters/Query-Name'
        - $ref: '#/components/parameters/Query-TagSet'
      responses:
        '200':
          description: Returns and array of objects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DB-Object'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
...
  /object/{objectId}/sync:
    get:
      summary: Synchronize an existing object
      description: >-
        Synchronize an existing COMS object so that any changes made externally
        (e.g. using an S3 client to upload a new version) are tracked by COMS.
        This endpoint does not guarantee immediately synchronized results.
        Synchronization latency will be affected by the remaining number of
        objects awaiting synchronization.
      tags:
        - Sync
      operationId: syncObject
      parameters:
        - $ref: '#/components/parameters/Path-ObjectId'
      responses:
        '202':
          description: Returns the number of objects that have been added to the queue.
          content:
            text/plain:
              schema:
                type: integer
                example: 5
                description: number of objects added to the queue
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Error'
...
  /object/metadata:
    get:
      summary: List metadata for objects
      description: >-
        Gets a list of metadata matching the given parameters. Multiple
        Key/Value metadata pairs can be provided in the query to narrow down
        results. If none are provided all metadata will be returned for each
        object.
      operationId: fetchMetadata
      tags:
        - Metadata
      parameters:
        - $ref: '#/components/parameters/Header-Metadata'
        - $ref: '#/components/parameters/Query-BucketId'
        - $ref: '#/components/parameters/Query-ObjectId'
      responses:
        '200':
          description: Returns an array objects with matching metadata key/value pairs.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - objectId
                    - metadata
                  properties:
                    objectId:
                      type: string
                      format: uuid
                      example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
                    metadata:
                      type: array
                      items:
                        $ref: '#/components/schemas/DB-MetadataKeyValue'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Error'
  /object/tagging:
    get:
      summary: List Tags for objects
      description: >-
        Gets a list of tags matching the given parameters. Multiple Key/Value
        metadata pairs can be provided in the query to narrow down results. If
        none are provided all tags will be returned for each object.
      operationId: fetchTags
      tags:
        - Tagging
      parameters:
        - $ref: '#/components/parameters/Query-BucketId'
        - $ref: '#/components/parameters/Query-ObjectId'
        - $ref: '#/components/parameters/Query-TagSet'
      responses:
        '200':
          description: Returns an array objects with matching key/value tag pairs.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - objectId
                    - tagset
                  properties:
                    objectId:
                      type: string
                      format: uuid
                      example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
                    tagset:
                      type: array
                      items:
                        $ref: '#/components/schemas/DB-TagKeyValue'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Error'
...
  /sync/status:
    get:
      summary: Get sync queue size
      description: Returns the number of objects currently awaiting processing by COMS.
      operationId: syncStatus
      tags:
        - Sync
      responses:
        '200':
          description: Returns the number of objects currently in the queue
          content:
            text/plain:
              schema:
                type: integer
                example: 5
                description: number of objects currently in the queue
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
...
  /version/metadata:
    get:
      summary: List metadata for versions
      description: >-
        Gets a metadata matching the given versions. Multiple Key/Value pairs
        can be provided in the headers to narrow down results. If none are
        provided the full set of metadata will be returned.
      operationId: fetchMetadataForVersion
      tags:
        - Metadata
      parameters:
        - $ref: '#/components/parameters/Header-Metadata'
        - in: query
          name: s3VersionId
          description: >-
            One or an array of version identifiers created in S3. This parameter
            is incompatible with `versionId`
          schema:
            oneOf:
              - $ref: '#/components/schemas/S3VersionId'
              - type: array
                items:
                  $ref: '#/components/schemas/S3VersionId'
        - in: query
          name: versionId
          description: >-
            One or an array of version identifiers created in COMS. This
            parameter is incompatible with `s3VersionId`
          schema:
            oneOf:
              - $ref: '#/components/schemas/VersionId'
              - type: array
                items:
                  $ref: '#/components/schemas/VersionId'
      responses:
        '200':
          description: Returns an array of versions with matching key/value pairs.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - versionId
                    - metadata
                  properties:
                    versionId:
                      type: string
                      format: uuid
                      example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
                    metadata:
                      type: array
                      items:
                        $ref: '#/components/schemas/DB-MetadataKeyValue'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Error'
  /version/tagging:
    get:
      summary: List tags for versions
      description: >-
        Gets matching tags for matching versions. Multiple Key/Value pairs can
        be provided in the headers to narrow down results. If none are provided
        the full set of tags will be returned.
      operationId: fetchTagsForVersion
      tags:
        - Tagging
      parameters:
        - $ref: '#/components/parameters/Query-TagSet'
        - in: query
          name: s3VersionId
          description: >-
            One or an array of version identifiers created in S3. This parameter
            is incompatible with `versionId`
          schema:
            oneOf:
              - $ref: '#/components/schemas/S3VersionId'
              - type: array
                items:
                  $ref: '#/components/schemas/S3VersionId'
        - in: query
          name: versionId
          description: >-
            One or an array of version identifiers created in COMS. This
            parameter is incompatible with `s3VersionId`
          schema:
            oneOf:
              - $ref: '#/components/schemas/VersionId'
              - type: array
                items:
                  $ref: '#/components/schemas/VersionId'
      responses:
        '200':
          description: Returns an array of versions with matching key/value pairs.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - versionId
                    - tagset
                  properties:
                    versionId:
                      type: string
                      format: uuid
                      example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
                    tagset:
                      type: array
                      items:
                        $ref: '#/components/schemas/DB-TagKeyValue'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        default:
          $ref: '#/components/responses/Error'
components:
...
  parameters:
    Header-ContentDisposition:
      in: header
      name: Content-Disposition
      required: true
      schema:
        type: string
        description: >-
          A valid RFC 6266 header. Must be of type `attachment` with either a
          `filename` or `filename*` parameter using the encoding defined in RFC
          8187 (normally UTF-8). For example, `skwc’әnjíłc.txt` must be
          represented as a url-encoded string
          `skwc%E2%80%99%D3%99nj%C3%AD%C5%82c.txt`.
        example: >-
          attachment; filename=bittercherrytree.txt;
          filename*=UTF-8''skwc%E2%80%99%D3%99nj%C3%AD%C5%82c.txt
    Header-ContentLength:
      in: header
      name: Content-Length
      required: true
      schema:
        type: integer
        description: A valid RFC 2616 header. Must be present for PUT requests.
        maximum: 5497558138880
        example: 3495
    Header-ContentType:
      in: header
      name: Content-Type
      schema:
        type: string
        description: A valid RFC 2045 MIME type
        default: application/octet-stream
        example: application/octet-stream
    Header-Metadata:
      in: header
      name: x-amz-meta-*
      description: >-
        An arbitrary metadata key/value pair. Must contain the x-amz-meta-
        prefix to be valid. Multiple metadata pairs can be defined. Keys must be
        unique, contain no whitespace, and will be converted to lowercase.
      schema:
        type: string
        minimum: 1
        example:
          - x-amz-meta-foo
          - x-amz-meta-bar
          - x-amz-meta-baz
    Path-BucketId:
      in: path
      name: bucketId
      description: Uuid of a bucket
      required: true
      schema:
        type: string
        format: uuid
        example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
    Path-ObjectId:
      in: path
      name: objectId
      description: Uuid of an object
      required: true
      schema:
        type: string
        format: uuid
        example: 48a65990-2e48-46b2-94eb-7f4fe13468ea
    Query-DeleteMarker:
      in: query
      name: deleteMarker
      description: Boolean on object version DeleteMarker status
      schema:
        type: boolean
        example: true
    Query-Latest:
      in: query
      name: latest
      description: Boolean on object version is latest
      schema:
        type: boolean
        example: true
    Query-Active:
      in: query
      name: active
      description: Boolean on active status
      schema:
        type: boolean
        example: true
...
    Query-MimeType:
      in: query
      name: mimeType
      description: The object MIME Type
      schema:
        type: string
        description: The object MIME Type
        default: application/octet-stream
        example: text/plain
    Query-ObjectId:
      in: query
      name: objectId
      description: Uuid or array of uuids representing the object
      schema:
        oneOf:
          - type: string
            format: uuid
            example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
          - type: array
            items:
              type: string
              format: uuid
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
    Query-BucketId:
      in: query
      name: bucketId
      description: Uuid or array of uuids representing the bucket
      schema:
        oneOf:
          - type: string
            format: uuid
            example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
          - type: array
            items:
              type: string
              format: uuid
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
...
    Query-Name:
      in: query
      name: name
      description: The filename of the object (typically the original filename)
      schema:
        type: string
        example: foobar.txt
...
    Query-Path:
      in: query
      name: path
      description: The canonical S3 path string of the object
      schema:
        type: string
        example: coms/env/ac246e31-c807-496c-bc93-cd8bc2f1b2b4
...
    Query-Public:
      in: query
      name: public
      description: Boolean on public status
      schema:
        type: boolean
        example: true
    Query-TagSet:
      in: query
      name: tagset[*]
      description: >-
        Tags for the object, defined as a Key/Value tag. The query must be
        formatted in deepObject style notation, where a tag-set made out of
        multiple tags would be encoded something similar to
        `tagset[x]=a&tagset[y]=b`. Only one value can exist for a given tag key.
      schema:
        $ref: '#/components/schemas/DB-TagKeyValue'
      style: deepObject
      explode: true
...
  schemas:
...
    DB-Metadata:
      title: DB Metadata
      type: object
      required:
        - metadata
      properties:
        metadata:
          type: object
          description: User-defined metadata
          example:
            color: red
            department: finance
    DB-MetadataKeyValue:
      title: DB Metadata Key Value
      type: object
      required:
        - key
        - value
      properties:
        key:
          type: string
          description: The metadata key
          example: x
        value:
          type: string
          description: The metadata value
          example: a
    DB-Object:
      title: DB Object
      type: object
      allOf:
        - type: object
          required:
            - id
            - path
            - public
            - active
          properties:
            id:
              type: string
              description: The primary identifier for this object
              format: uuid
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
            path:
              type: string
              description: The canonical S3 path string of the object
              example: coms/env/foobar.txt
            public:
              type: boolean
              description: Determines whether this object is publicly accessible
              default: false
              example: false
            active:
              type: boolean
              description: Determines whether this object is considered active
              default: true
              example: true
            bucketId:
              type: string
              description: The primary identifier for the bucket
              format: uuid
              example: c05c7650-5f48-4e51-bf17-762e8fc121a1
            name:
              type: string
              description: The filename of the original file uploaded
              example: foobar.txt
        - $ref: '#/components/schemas/DB-TimestampUserData'
...
    DB-TagKeyValue:
      title: DB Tag Key Value
      type: object
      required:
        - key
      properties:
        key:
          type: string
          description: The tag key
          example: x
        value:
          type: string
          description: The tag value
          example: a
    DB-Tags:
      title: DB Tags
      type: object
      allOf:
        - type: object
          properties:
            tags:
              type: object
              description: User-defined tags
              example:
                classification: Top Secret
                coms-id: 989de109-1487-423b-bf82-bee38c26ee3c
                priority: urgent
...
    Response-BadRequest:
      title: Response Bad Request
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            status:
              example: 400
            title:
              example: Bad Request
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/400
...
    Response-Conflict:
      title: Response Conflict
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            status:
              example: 409
            title:
              example: Conflict
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/409
...
    Response-Error:
      title: Response Error
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            status:
              example: 500
            title:
              example: Internal Server Error
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/500
    Response-Forbidden:
      title: Response Forbidden
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/403
            title:
              example: Forbidden
            status:
              example: 403
            detail:
              example: User lacks permission to complete this action
...
    Response-ObjectDetails:
      title: Response Object Details
      type: object
      allOf:
        - $ref: '#/components/schemas/DB-Object'
        - type: object
          required:
            - length
            - mimeType
            - Key
            - Location
          properties:
            length:
              type: integer
              description: Size of the file body in bytes
              maximum: 5497558138880
              example: 529
            mimeType:
              type: string
              description: The object MIME Type
              default: application/octet-stream
              example: text/plain
            versionId:
              type: string
              description: The primary identifier for this version in the COMS database
              format: uuid
              example: ac246e31-c807-496c-bc93-cd8bc2f1b2b4
        - $ref: '#/components/schemas/DB-Metadata'
        - $ref: '#/components/schemas/DB-Tags'
        - $ref: '#/components/schemas/S3-Object'
...
    Response-Unauthorized:
      title: Response Unauthorized
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/401
            title:
              example: Unauthorized
            status:
              example: 401
            detail:
              example: Invalid authorization credentials
...
    Response-ValidationError:
      title: Response Validation Error
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          required:
            - errors
          properties:
            errors:
              type: array
              items:
                type: object
                required:
                  - message
                properties:
                  value:
                    type: object
                    description: Contents of the field that was in error.
                    example: {}
                  message:
                    type: string
                    description: The error message for the field.
                    example: Invalid value `encoding`.
            status:
              example: 422
            title:
              example: Unprocessable Entity
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/422
...
    S3-Object:
      title: S3 Object
      type: object
      required:
        - $metadata
        - ETag
      properties:
        $metadata:
          $ref: '#/components/schemas/S3-Metadata'
        ETag:
          type: string
          description: S3 Identifier for a specific version of this object
          example: '"9d1aaa54b84e1d6ccc6e0477c5717fe3"'
        Bucket:
          type: string
          description: S3 Identifier for the bucket
          example: ezakyp
        Key:
          type: string
          description: Key of the object
          example: coms/env/00000000-0000-0000-0000-000000000000
        Location:
          type: string
          description: URL to the resource just uploaded
          example: >-
            https://your.objectstore.com/yourbucket/coms/env/00000000-0000-0000-0000-000000000000
        ServerSideEncryption:
          type: string
          description: >-
            If the object is stored using server-side encryption either with an
            AWS KMS key or an Amazon S3-managed encryption key, this field
            specifies the server-side encryption algorithm used when storing
            this object in Amazon S3 (for example, AES256, aws:kms).
          example: AES256
        s3VersionId:
          type: string
          description: A version identifier created in S3
          example: '1647462569641'
...
    S3VersionId:
      title: S3 Version ID
      type: string
      description: >-
        A version identifier created in S3. This parameter is incompatible with
        `versionId`
      maximum: 1024
      example: '1647462569641'
    VersionId:
      title: Version ID
      type: string
      description: >-
        A version identifier created by COMS. This parameter is incompatible
        with `s3VersionId`
      format: uuid
      example: c05c7650-5f48-4e51-bf17-762e8fc121a1
  responses:
    AddedQueueLength:
      description: Returns the number of objects that have been added to the queue
      content:
        text/plain:
          schema:
            type: integer
            example: 5
            description: Number of objects added to the queue
...
    BadRequest:
      description: Bad Request (Request is missing content or malformed)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-BadRequest'
...
    ContentTooLarge:
      description: Content Too Large
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-ContentTooLarge'
    Error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Error'
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Forbidden'
    LengthRequired:
      description: Length Required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-LengthRequired'
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Unauthorized'
    UnprocessableEntity:
      description: Unprocessable Content (Generally validation error(s))
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-ValidationError'
    UnsupportedMediaType:
      description: Unsupported Media Type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-UnsupportedMediaType'
  securitySchemes:
    BasicAuth:
      type: http
      description: Basic auth via environment defined username and password
      scheme: basic
    BearerAuth:
      type: http
      description: Bearer token auth using an OIDC issued JWT token
      scheme: bearer
      bearerFormat: JWT
    OpenID:
      type: openIdConnect
      description: OpenID Connect endpoint for acquiring JWT tokens
      openIdConnectUrl: >-
        https://logonproxy.gov.bc.ca/auth/realms/your-realm-name/.well-known/openid-configuration